% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.did_continuous}
\alias{print.did_continuous}
\title{print method for did_continuous}
\usage{
\method{print}{did_continuous}(x, ...)
}
\arguments{
\item{x}{A did_continuous object}

\item{...}{Undocumented}
}
\description{
A customized printed display for did_continous output
}
