% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_continuous.R
\name{did_continuous}
\alias{did_continuous}
\title{Main interface for did_continuous}
\usage{
did_continuous(
  df,
  Y,
  ID,
  T,
  D,
  Z = NULL,
  estimator = NULL,
  estimation_method = NULL,
  order = 1,
  noextrapolation = FALSE,
  placebo = NULL,
  weight = NULL,
  switchers = NULL,
  disaggregate = FALSE,
  aoss_vs_waoss = FALSE
)
}
\arguments{
\item{df}{df}

\item{Y}{Y}

\item{ID}{ID}

\item{T}{T}

\item{D}{D}

\item{Z}{Z}

\item{estimator}{estimator}

\item{estimation_method}{estimation_method Y}

\item{order}{order Y}

\item{noextrapolation}{noestrapolation}

\item{placebo}{placebo}

\item{weight}{weight Y}

\item{switchers}{switchers Y}

\item{disaggregate}{disaggregate Y}

\item{aoss_vs_waoss}{aoss_vs_waoss Y}
}
\description{
This program estimates the three estimators (aoss, waoss, iv-aoss) developped in de Chaisemartin, d'Haultfoeuille, Pasquier and Vazquez‐Bare,Difference-in-Differences Estimators for Treatments Continuously Distributed at Every Period (January 18, 2022).
}
